ADDED: shoot direction

+++ src/components/Position.hpp
@@ -15,9 +15,9 @@ namespace R_TYPE {
         /// @brief get the position 
         sf::Vector2f getPosition() {return (pos);};
 
-        void setX(int _x) {pos.x = _x;};
+        void setX(float _x) {pos.x = _x;};
 
-        void setY(int _y) {pos.y = _y;};
+        void setY(float _y) {pos.y = _y;};
 
         void operator=(const Position &other);
     private:
